requirements.txt
azure-functions
requests
redis
